+++ src/Stack/Options/NewParser.hs
@@ -23,9 +23,12 @@ newOptsParser = (,) <$> newOpts <*> initOptsParser
              help "Do not create a subdirectory for the project") <*>
         optional (templateNameArgument
             (metavar "TEMPLATE_NAME" <>
-             help "Name of a template or a local template in a file or a URL.\
-                  \ For example: foo or foo.hsfiles or ~/foo or\
-                  \ https://example.com/foo.hsfiles")) <*>
+             help "Name of a template - can take the form\
+                \ [[service:]username/]template with optional service name\
+                \ (github, gitlab, or bitbucket) \
+                \ and username for the service; or, a local filename such as\
+                \ foo.hsfiles or ~/foo; or, a full URL such as\
+                \ https://example.com/foo.hsfiles.")) <*>
         fmap
             M.fromList
             (many