+++ src/Stack/Types/TemplateName.hs
@@ -30,7 +30,20 @@ data TemplatePath = AbsPath (Path Abs File)
                   -- the template repository
                   | UrlPath String
                   -- ^ a full URL
+                  | RepoPath RepoTemplatePath
   deriving (Eq, Ord, Show)
+
+-- | Details for how to access a template from a remote repo.
+data RepoTemplatePath = RepoTemplatePath
+    { rtpService  :: RepoService
+    , rtpUser     :: Text
+    , rtpTemplate :: Text
+    }
+    deriving (Eq, Ord, Show)
+
+-- | Services from which templates can be retrieved from a repository.
+data RepoService = Github | Gitlab | Bitbucket
+    deriving (Eq, Ord, Show)
 
 instance FromJSON TemplateName where
     parseJSON = withText "TemplateName" $
@@ -79,12 +92,13 @@ parseTemplateNameFromString fname =
                                            $ asum (validParses prefix hsf orig)
     validParses prefix hsf orig =
         -- NOTE: order is important
-        [ TemplateName (T.pack orig) . UrlPath <$> (parseRequest orig *> Just orig)
+        [ TemplateName prefix        . RepoPath <$> parseRepoPath hsf
+        , TemplateName (T.pack orig) . UrlPath <$> (parseRequest orig *> Just orig)
         , TemplateName prefix        . AbsPath <$> parseAbsFile hsf
         , TemplateName prefix        . RelPath <$> parseRelFile hsf
         ]
     expected = "Expected a template like: foo or foo.hsfiles or\
-               \ https://example.com/foo.hsfiles"
+               \ https://example.com/foo.hsfiles or github:user/foo"
 
 -- | Make a template name.
 mkTemplateName :: String -> Q Exp
@@ -98,6 +112,11 @@ mkTemplateName s =
                           AbsPath (Path fp) -> [|AbsPath (Path fp)|]
                           RelPath (Path fp) -> [|RelPath (Path fp)|]
                           UrlPath fp -> [|UrlPath fp|]
+                          RepoPath (RepoTemplatePath sv u t) ->
+                            case sv of
+                                Github    -> [|RepoPath $ RepoTemplatePath Github u t|]
+                                Gitlab    -> [|RepoPath $ RepoTemplatePath Gitlab u t|]
+                                Bitbucket -> [|RepoPath $ RepoTemplatePath Bitbucket u t|]
 
 -- | Get a text representation of the template name.
 templateName :: TemplateName -> Text
@@ -106,3 +125,27 @@ templateName (TemplateName prefix _) = prefix
 -- | Get the path of the template.
 templatePath :: TemplateName -> TemplatePath
 templatePath (TemplateName _ fp) = fp
+
+defaultRepoUserForService :: RepoService -> Maybe Text
+defaultRepoUserForService Github = Just "commercialhaskell"
+defaultRepoUserForService _      = Nothing
+
+-- | Parses a template path of the form @github:user/template@.
+parseRepoPath :: String -> Maybe RepoTemplatePath
+parseRepoPath s =
+  case T.splitOn ":" (T.pack s) of
+    ["github"    , rest] -> parseRepoPathWithService Github rest
+    ["gitlab"    , rest] -> parseRepoPathWithService Gitlab rest
+    ["bitbucket" , rest] -> parseRepoPathWithService Bitbucket rest
+    _                    -> Nothing
+
+-- | Parses a template path of the form @user/template@, given a service
+parseRepoPathWithService :: RepoService -> Text -> Maybe RepoTemplatePath
+parseRepoPathWithService service path =
+  case T.splitOn "/" path of
+    [user, name] -> Just $ RepoTemplatePath service user name
+    [name]       -> do
+        repoUser <- defaultRepoUserForService service
+        Just $ RepoTemplatePath service repoUser name
+    _            -> Nothing
+